Revert r12117, thus re-opening issue #2144.
We're doing this because the fixing of issue #2144 causes issue #2148
to be revealed, which is far nastier.  Issue #2148 isn't likely to be fixed
anytime soon;  it has its own branch, and requires a wc-format
change (and thus a 1.X release).

Therefore, both issues #2144 and #2148 need to be fixed as a pair, and
are currently assigned to the 1.3 release.  See the issues for details.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@853883 13f79535-47bb-0310-9956-ffa450edef68