Bugfix:  make copy_tests.py #27 pass over ra_dav.
When ra_dav does a PROPFIND on a path like 'file1/file2', it's
expecting a 404 'not found' error back, so that can retry the
operation on a parent.  But instead, mod_dav_svn was throwing a 500
when libsvn_fs returned SVN_ERR_FS_NOT_DIRECTORY.  The solution is to
have mod_dav_svn trap this specific error and still return 404.

* subversion/mod_dav_svn/repos.c 
  (dav_svn__fs_check_path):  new helper func to wrap svn_fs_check_path().
  (dav_svn_prep_regular,
   dav_svn_prep_working,
   dav_svn_resource_kind):   call the wrapper.



git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@853918 13f79535-47bb-0310-9956-ffa450edef68