* packages/windows-innosetup: Renamed from packages/win32-innosetup.
The win32 part of the name is meaningless. Inno Setup doesn't care if it's 32-
or 64-bit and handles both systems without problems.

The Setup System is based on that the setup maker person is using a Working
Copy for each setup series (trunk, 1.1.x series, 1.2.x series and so on) not
different win*-innosetup folders in the repository.



git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@854255 13f79535-47bb-0310-9956-ffa450edef68