On argument-parsing errors, don't print help, just an error message. The errors in question are mainly for too few or too many arguments.
Printing help may be more helpful to a novice in the short term, but in the
long term it gives a false sense of security, causing users to develop a bad
habit (in that not all commands print help when invoked without arguments).
Also, when invoking the command accidentally without arguments (perhaps
giving as an argument an environment variable which was accidentally
undefined), seeing a long help message is annoying.  Also it makes a
compatibility problem if we want to extend a command to have an argument-free
form in the future.

* subversion/clients/cmdline/main.c (main):
  Don't replace a generic argument-parsing error with help.

* subversion/clients/cmdline/merge-cmd.c (svn_cl__merge):
* subversion/svnadmin/main.c (main):
* subversion/svndumpfilter/main.c (main):
* subversion/svnlook/main.c (main):
  Don't augment argument-parsing errors with help.  Be more explicit about
  whether there are too few or too many arguments, in line with r17336.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@857423 13f79535-47bb-0310-9956-ffa450edef68