Fix issue #2442, verbose list broken with peg revisions over ra_dav.
It turns out that the underlying issue here is a difference in the way
that svn_ra_get_locks worked in ra_dav versus the other ra layers.  In
both ra_local and ra_svn it's perfectly valid to give svn_ra_get_locks
a nonexistent path, it'll just return no locks for it, but ra_dav says
that's an error and returns path not found.

Since this does seem to cause a problem in a reasonable case, I've made
ra_dav act like the other RA layers, and just return no locks in that
situation.  Because this is kind of strange behavior compared to other
commands, I've noted it in the doxygen comments for svn_ra_get_locks.

* subversion/include/svn_ra.h
  (svn_ra_get_locks): Note behavior when path refers to a nonexistent
   location.

* subversion/libsvn_ra_dav/fetch.c
  (svn_ra_dav__get_locks): Don't return an error when the path doesn't
   exist in HEAD.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@857559 13f79535-47bb-0310-9956-ffa450edef68