Replace the public SVN_STREAM_CHUNK_SIZE constant with an internal one that we can tweak at will.  Use a 16K chunk size instead of 100K.
The reason for this change is that using 100K buffers doesn't buy us anything
performance-wise.  It also decreases responsiveness in some situations
(for example, you can't cancel an svn cat command between 100K chunks).
See this thread for an indication that a power of 2 actually may increase
performance on a particular platform:
http://svn.haxx.se/dev/archive-2004-11/0123.shtml

Also, note that this value isn't meant to be the optimal buffer size
for file I/O.  Files have their own buffering if needed.

Suggested by: ringstrom
(A long time ago.)

* subversion/include/svn_types.h (SVN_STREAM_CHUNK_SIZE): Deprecate.
  (SVN__STREAM_CHUNK_SIZE): New *internal* macro.  All uses of
  SVN_STREAM_CHUNK_SIZE replace with this macro throughout the code.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@857671 13f79535-47bb-0310-9956-ffa450edef68