Avoid breaking working copies with a switch that changes the repository root part of the URL.
This has two effects:
a) It disallows switch from changing the repository root part of the URLs.
b) It does not install a repos entry field during an update if the 
repository root URL would not be an ancestor path of the entry's URL.  This
avoids breaking inconsistent working copies that did not have a repos
attribute (i.e. that were checked out with a pre-1.3 client.)

Both the above cases result in corrupt working copies.

Suggested by: rooneg
              jrepenning

* subversion/include/svn_error_codes.h
  (SVN_ERR_WC_INVALID_SWITCH): New error code.

* subversion/libsvn_wc/update_editor.c
  (prep_directory): Only give repos root to svn_wc_ensure_adm2() if it is
  an ancestor path of the entry's URL.
  (open_root, open_directory): Only set repos root field of the entry if it
  actually is an ancestor path of the new URL.
  (make_editor): If there is a switch_url and we know the repository root,
  don't allow the switch to change the repository root of the URL.

* subversion/tests/svn/switch_tests.py (switch_change_repos_root): New test.
  (test_list): Run it.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@857739 13f79535-47bb-0310-9956-ffa450edef68