OS400/EBCDIC Port: Create new files with CCSID 1208.
This is one of several patches to allow Subversion to run on IBM's
OS400 V5R4.  It tags files created by Subversion with the CCSID 1208
(UTF-8), which is always correct for text files and irrelevant for
binary files, and allows OS400 applications and utilities to operate
correctly on these files.

Approved By: Philip Martin <philip@codematters.co.uk>
             Julian Foad <julianfoad@btopenworld.com>
             
* subversion/libsvn_subr/io.c
   (file_open): new function.
   (svn_io_open_unique_file2, svn_io_file_open): Replace calls to
    apr_file_open with new function.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@858662 13f79535-47bb-0310-9956-ffa450edef68