This time for sure!
Because we can have files and directories interspersed with each other in
the editor drive the old way of managing the file pools for ra_dav's replay
was broken, and could result in reusing destroyed pools, which is bad.  Now
we have per-directory file pools, so this is not a problem anymore.

* subversion/libsvn_ra_dav/replay.c
  (replay_baton_t): Remove the file pool.
  (dir_item_t): Add a per-dir file pool.
  (push_dir): Initialize the new per-directory file pool.
  (start_element, end_element): Adjust to having a per-directory file pool.

* subversion/tests/cmdline/svnsync_tests.py
  (file_dir_file): New test.
  (test_list): Run new test.

* subversion/tests/cmdline/svnsync_tests_data/file-dir-file.dump: New dump
  file for testing.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@859456 13f79535-47bb-0310-9956-ffa450edef68