As discussed on the svnmerge mailing list, there is no actual reason to block a merge just because there are pending updates on the repository. A merge does not do any commit, it just modifies the working copy. And it is normal to perform updates before committing. Also, the check is actually meaningless, because by the time the merge is finished, it might well be that there are new pending updates on the repository.
It makes sense, instead, to make sure the working copy is clean, so not to
mistakenly commit unrelated modifies together with a merge. This check can
be disabled with --force.

A nice side effect of this patch is that it speeds out merge operations as
one less server/client round-trip is required.

 * contrib/client-side/svnmerge.py:
   (check_dir_clean): Do not check for pending updates.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@859720 13f79535-47bb-0310-9956-ffa450edef68