Make the libsvn_fs_fs handling of buffer sizes consistent.  Previous to this change some parts of the code used one size buffer for reading the contents of revision files, but others used a much smaller size buffer.
This change centers around the idea that we want to allow path lengths
up to 4096 bytes, so buffers are sized such that they can hold a 4096 
byte path, plus whatever goes on the same line as a path.

This is part of work on issue 2508.  It doesn't "fix" things, as you
can still hit the max size of the buffer with sufficient determination,
but it does make the limit larger than the maximum path length on most
platforms, so you really have to go out of your way to hit it.

* subversion/libsvn_fs_fs/fs_fs.c
  (FSFS_MAX_PATH_LEN, MAX_HEADERS_STR_LEN):  New constants.
  (read_header_block, read_change): Use MAX_HEADER_STR_LEN to size our
   input buffer.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@860172 13f79535-47bb-0310-9956-ffa450edef68