Rename a parameter of svn_fs_open() from 'config' to 'fs_config'. The usage of 'config' was inconsistent with:  - the function doc-comment  - the actual implementation of the function  - the conventional name for that kind of hash elsewhere in the svn_fs API.
Also, 'apr_hash_t *config' is conventionally a hash of a completely different
type, elsewhere in our API.

Incidentally, this renaming also causes the SWIG bindings to stop applying
erroneous typemaps to this parameter.

I believe it is permitted to rename a parameter, even in a frozen API - if you
think/know otherwise, please speak up.

* subversion/include/svn_fs.h (svn_fs_open): Rename parameter 3 to 'fs_config'.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@860529 13f79535-47bb-0310-9956-ffa450edef68